# doc/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "WellPaw API is a RESTful API for managing and tracking pet health",
        "title": "WellPaw API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:50001",
    "basePath": "/api",
    "paths": {
...
        "/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get user all info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get User All Info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/user/notification/calendar": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "manage calendar notification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Manage Calendar Notification",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/user/notification/food": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "manage food notification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Manage Food Notification",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/user/paymentintent/{payment_intent_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get payment intent by id to use retry payment in frontend (in case of payment failed)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get Payment Intent By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment intent id",
                        "name": "payment_intent_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.PaymentIntentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/user/paymentmethod": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "update payment method",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update Payment Method",
                "parameters": [
                    {
                        "description": "Update payment method payload",
                        "name": "UpdatePaymentMethodPayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.PaymentMethodUpdatePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/user/subscription": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get all subscriptions plan and their details provide for customer to select",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get All Subscriptions Plan",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.SubscriptionPlanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/user/subscription/cancel/{subscription_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "cancel subscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Cancel Subscription",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.SubscriptionHistoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/user/subscription/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get all subscription histories by customer id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get All Subscriptions By Customer ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Last id",
                        "name": "last_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.SubscriptionHistoryPaginationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/user/subscription/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "start subscription after attached payment method and return payment intent detail to confirm payment in frontend",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Start Subscription",
                "parameters": [
                    {
                        "description": "Start subscription payload",
                        "name": "StartSubscriptionPayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.StartSubscriptionPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.PaymentIntentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/user/subscription/update": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "update subscription return payment intent detail to confirm payment in frontend",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update Subscription",
                "parameters": [
                    {
                        "description": "Update subscription payload",
                        "name": "UpdateSubscriptionPayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.UpdateSubscriptionPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.PaymentIntentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.HTTPResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_231031_wellpaw-backend_internal_model.HTTPResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_231031_wellpaw-backend_internal_model.PaymentIntentResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.PaymentInvoice"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "github_com_231031_wellpaw-backend_internal_model.PaymentInvoice": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "client_secret": {
                    "type": "string"
                },
                "default_payment_method_id": {
                    "type": "string"
                },
                "payment_intent_status": {
                    "type": "string"
                },
                "subscription_status": {
                    "type": "string"
                }
            }
        },
        "github_com_231031_wellpaw-backend_internal_model.PaymentMethodUpdatePayload": {
            "type": "object",
            "properties": {
                "payment_method_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_231031_wellpaw-backend_internal_model.StartSubscriptionPayload": {
            "type": "object",
            "properties": {
                "payment_method_id": {
                    "type": "string"
                },
                "subscription_plan_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_231031_wellpaw-backend_internal_model.SubscriptionHistoryPagination": {
            "type": "object",
            "properties": {
                "last_id": {
                    "type": "string"
                },
                "subscriptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.SubscriptionHistory"
                    }
                }
            }
        },
        "github_com_231031_wellpaw-backend_internal_model.SubscriptionHistoryPaginationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.SubscriptionHistoryPagination"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "github_com_231031_wellpaw-backend_internal_model.SubscriptionHistoryResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.SubscriptionHistory"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "github_com_231031_wellpaw-backend_internal_model.SubscriptionPlan": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                },
                "features": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "interval": {
                    "type": "string"
                },
                "interval_count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_231031_wellpaw-backend_internal_model.SubscriptionPlanResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.SubscriptionPlan"
                    }
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_231031_wellpaw-backend_internal_model.TierType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "FREE",
                "WEEKS",
                "MONTHS",
                "YEARS"
            ]
        },
...
        "github_com_231031_wellpaw-backend_internal_model.UpdateSubscriptionPayload": {
            "type": "object",
            "properties": {
                "new_subscription_plan_id": {
                    "type": "string"
                }
            }
        },
        "github_com_231031_wellpaw-backend_internal_model.User": {
            "type": "object",
            "properties": {
                "bcs_free": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "string"
                },
                "device_token": {
                    "type": "string"
                },
                "disease_free": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "food_free": {
                    "type": "integer"
                },
                "food_plan_free": {
                    "type": "integer"
                },
                "foods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.Food"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "noti_calendars": {
                    "type": "boolean"
                },
                "noti_food": {
                    "type": "boolean"
                },
                "password": {
                    "type": "string"
                },
                "payment_method_id": {
                    "type": "string"
                },
                "payment_plan": {
                    "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.TierType"
                },
                "payments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.Payment"
                    }
                },
                "pets": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.Pet"
                    }
                },
                "profile_free": {
                    "type": "integer"
                },
                "subscription_status": {
                    "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.SubscriptionStatusType"
                },
                "tier": {
                    "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.TierType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_231031_wellpaw-backend_internal_model.UserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_231031_wellpaw-backend_internal_model.User"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}